requirements.txt
colorzero==1.1
gpiozero==1.5.1
pkg-resources==0.0.0
RPi.bme280==0.2.3
RPi.GPIO==0.7.0
smbus2==0.3.0
